backend/src/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
coverage==7.10.5
Django==5.2.5
djangorestframework==3.16.1
flake8==7.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
sqlparse==0.5.3
